+++ src/types/common.ts
@@ -11,29 +11,3 @@ export interface ButtonModalProps {
   isOpen: boolean;
   setIsOpen: Dispatch<SetStateAction<boolean>>;
 }
-
-export interface IAnswers {
-  0: string[] | null;
-  1: string | null;
-  2: string | null;
-  3: string | null;
-  4: string | null;
-  5: string | null;
-}
-export interface IBubbleSectionProps {
-  step: 0 | 1 | 2 | 3 | 4 | 5;
-  answers: IAnswers;
-  setAnswers: Dispatch<SetStateAction<IAnswers>>;
-  moveToNextStep: () => void;
-  pbStation?: {
-    branchName: string;
-    branchAddress: string;
-    branchLatitude: number;
-    branchLongitude: number;
-  };
-  consultTime?: {
-    consultEnd: string;
-    consultStart: string;
-    notice: string;
-  };
-}
